Change get_color return type

+++ simulator/body.h
@@ -28,7 +28,7 @@ class Body
     double get_x_momentum(); //return x momentum
     double get_y_momentum(); //return y momentum
     double get_orbital_momentum(); //return orbital angular momentum
-    double get_color();
+    int get_color();
     double get_total_energies();
     
     //variables
